requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.21.0
matplotlib>=3.4.0
Pillow>=8.3.0
tqdm>=4.62.0
scikit-learn>=1.0.0
ultralytics>=8.0.0
pandas>=1.3.0
PyYAML>=6.0
mlflow>=2.0.0 